Fixed bugs; finished some features, store file system and its data; finished javadoc for the moment

+++ src/r0p3/FileData.java
@@ -4,8 +4,8 @@ public class FileData {
 
     private String extention = "";
     private Integer number;
-    private Double size;
-    private Double percentage = 0.0;
+    private Long size;
+    private Long percentage = (long)0;
 
     /**
      * DataFile constructor
@@ -16,7 +16,7 @@ public class FileData {
         String[] aux = extention.split(".");
         this.number = 1;
         this.extention = "." + aux[aux.length - 1];
-        this.size = 0.0;
+        this.size = (long)0;
     }
 
 
@@ -25,9 +25,9 @@ public class FileData {
      * It requires a file extention and a size
      *
      * @param   extention   a String which represents the file extention
-     * @param   size        a Double which represents the file size
+     * @param   size        a Long which represents the file size
      * */
-    public FileData (String extention, Double size) {
+    public FileData (String extention, Long size) {
         String[] aux = extention.split(".");
         this.number = 1;
         this.extention = "." + aux[aux.length - 1];
@@ -46,14 +46,19 @@ public class FileData {
 
 
     /**
+     * Number of files scanned getter
      *
+     * @return  number of files scanned
      * */
     public Integer getNumberOfFiles () {
         return number;
     }
 
     /**
+     * Add some files, if they were scanned
+     * It helps because java hashmaps are miserable
      *
+     * @param   n   number of files to add
      * */
     public void addFileNumber (Integer n) {
         number += n;
@@ -63,37 +68,37 @@ public class FileData {
     /**
      * Add a size to that file extention
      *
-     * @param   s   a Double which represents the file size
+     * @param   s   a Long which represents the file size
      * */
-    public void addSize (Double s) {
+    public void addSize (Long s) {
         size += s;
     }
 
     /**
      * File size getter
      *
-     * @return  a Double which represents the file size
+     * @return  a Long which represents the file size
      * */
-    public Double getSize () {
+    public Long getSize () {
         return size;
     }
 
     /**
      * Calculate the percentage that represents the total presence of the file with this extension
      *
-     * @param   s       a Double which represents the file size
-     * @param   total   a Double which represents the total disk size
+     * @param   s       a Long which represents the file size
+     * @param   total   a Long which represents the total disk size
      * */
-    public void calculatePercentage (Double s, Double total) {
+    public void calculatePercentage (Long s, Long total) {
         percentage = s * 100 / total;   // regla de 3, ole ole
     }
 
     /**
      * File percentage getter
      *
-     * @return  a Double which represents the file percentage
+     * @return  a Long which represents the file percentage
      * */
-    public Double getPercentage () {
+    public Long getPercentage () {
         return percentage;
     }
